Module phaselock::data

Provides types useful for representing PhaseLock’s data structures

This module provides types for representing consensus internal state, such as the Leaf, PhaseLock’s version of a block, and the QuorumCertificate, representing the threshold signatures fundamental to consensus.

Structs

BlockHash

External wrapper type

Leaf

A node in PhaseLock’s consensus-internal merkle tree.

LeafHash

External wrapper type

QuorumCertificate

The type used for Quorum Certificates

StateHash

External wrapper type

TransactionHash

External wrapper type

VecQuorumCertificate

QuorumCertificate variant using a Vec rather than a const-generic array

VerifyHash

External wrapper type

Enums

Stage

Represents the stages of consensus

Functions

create_verify_hash

This concatenates the encoding of leaf_hash, view, and stage, in that order, and hashes the result.