Enum phaselock::PhaseLockError

#[non_exhaustive]
pub enum PhaseLockError {
    FailedToMessageLeader {
        stage: Stage,
        source: NetworkError,
    },
    FailedToBroadcast {
        stage: Stage,
        source: NetworkError,
    },
    BadOrForgedQC {
        stage: Stage,
        bad_qc: VecQuorumCertificate,
    },
    FailedToAssembleQC {
        stage: Stage,
        source: Error,
    },
    BadBlock {
        stage: Stage,
    },
    InconsistentBlock {
        stage: Stage,
    },
    NetworkFault {
        source: NetworkError,
    },
    ItemNotFound {
        hash: Vec<u8, Global>,
    },
    StorageError {
        source: StorageError,
    },
    InvalidState {
        context: String,
    },
    Continue,
}

Reexport error type Error type for PhaseLock

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

FailedToMessageLeader

Fields

stage: Stage

The stage the failure occurred in

source: NetworkError

The underlying network fault

Failed to Message the leader in the given stage

FailedToBroadcast

Fields

stage: Stage

The stage the failure occurred in

source: NetworkError

The underlying network fault

Failed to broadcast a message on the network

BadOrForgedQC

Fields

stage: Stage

The stage the failure occurred in

bad_qc: VecQuorumCertificate

The bad quorum certificate

Bad or forged quorum certificate

FailedToAssembleQC

Fields

stage: Stage

The stage the error occurred in

source: Error

The underlying crypto fault

Failed to assemble a quorum certificate

BadBlock

Fields

stage: Stage

The stage the error occurred in

A block failed verification

InconsistentBlock

Fields

stage: Stage

The stage the error occurred in

A block was not consistent with the existing state

NetworkFault

Fields

source: NetworkError

Underlying network fault

Failure in networking layer

ItemNotFound

Fields

hash: Vec<u8, Global>

Hash of the missing item

Item was not present in storage

StorageError

Fields

source: StorageError

Underlying error

Error accesing storage

InvalidState

Fields

context: String

Context

Invalid state machine state

Continue

Internal value used to drive the state machine

Implementations

impl PhaseLockError

Reexport error type

pub fn get_stage(&self) -> Option<Stage>

Returns the stage this error happened in, if such information exists