Module phaselock::traits

Contains traits consumed by PhaseLock

Re-exports

pub use election::Election;

Modules

election

Sortition trait

implementations

Module for publicly usable implementations of the traits

Enums

NetworkError

Error type for networking

Traits

BlockContents

Abstraction over the contents of a block

NetworkReliability

interface describing how reliable the network is

NetworkingImplementation

Describes, generically, the behaviors a networking implementation must have

NodeImplementation

Node implementation aggregate trait

State

Abstraction over the state that blocks modify

StatefulHandler

Trait for a stateful event handler

Storage

Abstraction over on disk persistence of node state

Type Definitions

BoxedFuture

A boxed future trait object with a static lifetime

StorageResult

Result for a storage type