Trait phaselock::traits::election::Election

pub trait Election<const N: usize> {
    type StakeTable;
    type SelectionThreshold;
    type State;
    type VoteToken;
    type ValidatedVoteToken;
    fn get_stake_table(&self, state: &Self::State) -> Self::StakeTable;

    fn get_leader(&self, table: &Self::StakeTable, view_number: u64) -> PubKey;

    fn get_votes(
        &self, 
        table: &Self::StakeTable, 
        selection_threshold: Self::SelectionThreshold, 
        view_number: u64, 
        pub_key: PubKey, 
        token: Self::VoteToken, 
        next_state: StateHash<N>
    ) -> Option<Self::ValidatedVoteToken>;

    fn get_vote_count(&self, token: &Self::ValidatedVoteToken) -> u64;

    fn make_vote_token(
        &self, 
        table: &Self::StakeTable, 
        selection_threshold: Self::SelectionThreshold, 
        view_number: u64, 
        private_key: &PrivKey, 
        next_state: StateHash<N>
    ) -> Option<Self::VoteToken>;
}

Describes how PhaseLock chooses committees and leaders

Associated Types

type StakeTable

Data structure describing the currently valid states

type SelectionThreshold

The threshold for membership selection.

type State

The state type this election implementation is bound to

type VoteToken

A membership proof

type ValidatedVoteToken

A type stated, validated membership proof

Required methods

fn get_stake_table(&self, state: &Self::State) -> Self::StakeTable

Returns the table from the current committed state

fn get_leader(&self, table: &Self::StakeTable, view_number: u64) -> PubKey

Returns leader for the current view number, given the current stake table

fn get_votes(
&self,
table: &Self::StakeTable,
selection_threshold: Self::SelectionThreshold,
view_number: u64,
pub_key: PubKey,
token: Self::VoteToken,
next_state: StateHash<N>
) -> Option<Self::ValidatedVoteToken>

Validates a vote token and returns the number of seats that it has

Salt: Hash of the leaf that is being proposed

fn get_vote_count(&self, token: &Self::ValidatedVoteToken) -> u64

Returns the number of votes the validated vote token has

fn make_vote_token(
&self,
table: &Self::StakeTable,
selection_threshold: Self::SelectionThreshold,
view_number: u64,
private_key: &PrivKey,
next_state: StateHash<N>
) -> Option<Self::VoteToken>

Attempts to generate a vote token for self

Returns None if the number of seats would be zero