Enum phaselock::traits::NetworkError

pub enum NetworkError {

    ListenerSend,
    CouldNotDeliver,
    NoSuchNode,
    FailedToSerialize {
        source: Error,
    },
    FailedToDeserialize {
        source: Error,
    },
    WebSocket {
        source: Error,
    },
    ExecutorError {
        source: Error,
    },
    SocketDecodeError {
        input: String,
        source: Error,
    },
    FailedToBindListener {
        source: Error,
    },
    NoSocketsError {
        input: String,
    },
    Other {
        inner: Box<dyn Error + Send + Sync>,
    },
    ChannelSend,
    IdentityHandshake,
    ShutDown,

}

Error type for networking

Variants

ListenerSend

A Listener failed to send a message

CouldNotDeliver

Could not deliver a message to a specified recipient

NoSuchNode

Attempted to deliver a message to an unknown node

FailedToSerialize

Failed to serialize a message

Fields of FailedToSerialize

source: Error

Originating bincode error

FailedToDeserialize

Failed to deserealize a message

Fields of FailedToDeserialize

source: Error

originating bincode error

WebSocket

WebSockets specific error

Fields of WebSocket

source: Error

Originating websockets error

ExecutorError

Error orginiating from within the executor

Fields of ExecutorError

source: Error

Originating async_std error

SocketDecodeError

Failed to decode a socket specification

Fields of SocketDecodeError

input: String

Input that was given

source: Error

Originating io error

FailedToBindListener

Failed to bind a listener socket

Fields of FailedToBindListener

source: Error

originating io error

NoSocketsError

No sockets were open

Fields of NoSocketsError

input: String

Input that was given

Other

Generic error type for compatibility if needed

Fields of Other

inner: Box<dyn Error + Send + Sync>

Originating error

ChannelSend

Channel error

IdentityHandshake

Could not complete handshake

ShutDown

The underlying connection has been shut down