Struct phaselock::traits::implementations::MemoryNetwork

pub struct MemoryNetwork<T> { /* fields omitted */ }

In memory only network simulator.

This provides an in memory simulation of a networking implementation, allowing nodes running on the same machine to mock networking while testing other functionality.

Under the hood, this simply maintains mpmc channels to every other MemoryNetwork insane of the same group.

Implementations

impl<T> MemoryNetwork<T> where
T: Clone + Serialize + DeserializeOwned + Send + Sync + Debug + 'static,

pub fn new(
pub_key: PubKey,
master_map: Arc<MasterMap<T>>,
reliability_config: Option<impl 'static + NetworkReliability>
) -> MemoryNetwork<T>

Creates a new MemoryNetwork and hooks it up to the group through the provided MasterMap