Trait phaselock::traits::BlockContents

pub trait BlockContents<const N: usize>: Serialize + DeserializeOwned + Clone + Debug + Hash + PartialEq + Eq + Send + Sync + Unpin {
    type Error: Error + Debug + Send + Sync;
    type Transaction: Clone + Serialize + DeserializeOwned + Debug + Hash + PartialEq + Eq + Sync + Send;
    fn add_transaction_raw(
        &self, 
        tx: &Self::Transaction
    ) -> Result<Self, Self::Error>;

    fn hash(&self) -> BlockHash<N>;

    fn hash_transaction(tx: &Self::Transaction) -> BlockHash<N>;

    fn hash_bytes(bytes: &[u8]) -> BlockHash<N>;
}

Abstraction over the contents of a block

This trait encapsulates the behaviors that a block must have in order to be used by consensus:

Must have a predefined error type (BlockContents::Error)

Must have a transaction type that can be compared for equality, serialized and serialized, sent between threads, and can have a hash produced of it (hash_transaction)

Must be able to be produced incrementally by appending transactions (add_transaction_raw)

Must be hashable (hash)

Associated Types

type Error: Error + Debug + Send + Sync

The error type for this type of block

type Transaction: Clone + Serialize + DeserializeOwned + Debug + Hash + PartialEq + Eq + Sync + Send

The type of the transitions we are applying

Required methods

fn add_transaction_raw(
&self,
tx: &Self::Transaction
) -> Result<Self, Self::Error>

Attempts to add a transaction, returning an Error if it would result in a structurally invalid block

Errors

Should return an error if this transaction leads to an invalid block

fn hash(&self) -> BlockHash<N>

Produces a hash for the contents of the block

fn hash_transaction(tx: &Self::Transaction) -> BlockHash<N>

Produces a hash for a transaction

TODO: Abstract out into transaction trait

fn hash_bytes(bytes: &[u8]) -> BlockHash<N>

Produces a hash for an arbitrary sequence of bytes

Used to produce hashes for internal PhaseLock control structures