Trait phaselock::traits::NetworkingImplementation

pub trait NetworkingImplementation<M>: Send + Sync where
    M: Serialize + DeserializeOwned + Send + Clone + 'static, {
    fn broadcast_message(
        &self, 
        message: M
    ) -> BoxFuture<'_, Result<(), NetworkError>>;

    fn message_node(
        &self, 
        message: M, 
        recipient: PubKey
    ) -> BoxFuture<'_, Result<(), NetworkError>>;

    fn broadcast_queue(&self) -> BoxFuture<'_, Result<Vec<M>, NetworkError>>;

    fn next_broadcast(&self) -> BoxFuture<'_, Result<M, NetworkError>>;

    fn direct_queue(&self) -> BoxFuture<'_, Result<Vec<M>, NetworkError>>;

    fn next_direct(&self) -> BoxFuture<'_, Result<M, NetworkError>>;

    fn known_nodes(&self) -> BoxFuture<'_, Vec<PubKey>>;

    fn obj_clone(&self) -> Box<dyn NetworkingImplementation<M> + 'static>;
}

Describes, generically, the behaviors a networking implementation must have

Required methods

fn broadcast_message(
&self,
message: M
) -> BoxFuture<'_, Result<(), NetworkError>>

Broadcasts a message to the network

Should provide that the message eventually reach all non-faulty nodes

fn message_node(
&self,
message: M,
recipient: PubKey
) -> BoxFuture<'_, Result<(), NetworkError>>

Sends a direct message to a specific node

fn broadcast_queue(&self) -> BoxFuture<'_, Result<Vec<M>, NetworkError>>

Moves out the entire queue of received broadcast messages, should there be any

Provided as a future to allow the backend to do async locking

fn next_broadcast(&self) -> BoxFuture<'_, Result<M, NetworkError>>

Provides a future for the next received broadcast

Will unwrap the underlying NetworkMessage

fn direct_queue(&self) -> BoxFuture<'_, Result<Vec<M>, NetworkError>>

Moves out the entire queue of received direct messages to this node

fn next_direct(&self) -> BoxFuture<'_, Result<M, NetworkError>>

Provides a future for the next received direct message to this node

Will unwrap the underlying NetworkMessage

fn known_nodes(&self) -> BoxFuture<'_, Vec<PubKey>>

Node’s currently known to the networking implementation

Kludge function to work around leader election

fn obj_clone(&self) -> Box<dyn NetworkingImplementation<M> + 'static>

Object safe clone