Trait phaselock::traits::State

pub trait State<const N: usize>: Serialize + DeserializeOwned + Clone + Debug + Hash + PartialEq<Self> + Eq + Send + Sync + Unpin {
    type Error: Error + Debug + Send + Sync;
    type Block: BlockContents<N>;
    fn next_block(&self) -> Self::Block;

    fn validate_block(&self, block: &Self::Block) -> bool;

    fn append(&self, block: &Self::Block) -> Result<Self, Self::Error>;

    fn on_commit(&self);
}

Abstraction over the state that blocks modify

This trait represents the behaviors that the ‘global’ ledger state must have:

A defined error type (Error)

The type of block that modifies this type of state (Block)

A method to get a template (empty) next block from the current state (next_block)

The ability to validate that a block is actually a valid extension of this state (validate_block)

The ability to produce a new state, with the modifications from the block applied (append)

Associated Types

type Error: Error + Debug + Send + Sync

The error type for this particular type of ledger state

type Block: BlockContents<N>

The type of block this state is associated with

Required methods

fn next_block(&self) -> Self::Block

Returns an empty, template next block given this current state

fn validate_block(&self, block: &Self::Block) -> bool

Returns true if and only if the provided block is valid and can extend this state

fn append(&self, block: &Self::Block) -> Result<Self, Self::Error>

Appends the given block to this state, returning an new state

Errors

Should produce and error if appending this block would lead to an invalid state

fn on_commit(&self)

Gets called to notify the persistence backend that this state has been committed