Trait phaselock::traits::StatefulHandler

pub trait StatefulHandler<const N: usize>: 'static + Send + Sync + Debug {
    type Block: 'static + BlockContents<N>;
    type State: State<N>;
    fn notify(
        &mut self, 
        blocks: Vec<Self::Block, Global>, 
        states: Vec<Self::State, Global>
    );
}

Trait for a stateful event handler

The PhaseLock instance will keep around the provided value of this type, and call the notify method every time a series of blocks are committed.

A do-nothing implementation (Stateless) is provided, as a convince for implementations that do not need this functionality.

Associated Types

type Block: 'static + BlockContents<N>

Block type for this consensus implementation

type State: State<N>

State type for this consensus implementation

Required methods

fn notify(
&mut self,
blocks: Vec<Self::Block, Global>,
states: Vec<Self::State, Global>
)

The PhaseLock implementation will call this method, with the series of blocks and states that are being committed, whenever a commit action takes place.

The provided states and blocks are guaranteed to be in ascending order of age (newest to oldest).

This functionality is discretionary, and whatever behavior is implemented in this method should not impact consensus logic.