Trait phaselock::traits::Storage

pub trait Storage<B: BlockContents<N> + 'static, S: State<N, Block = B> + 'static, const N: usize>: Clone + Send + Sync {
    fn get_block<'b, 'a: 'b>(
        &'a self, 
        hash: &'b BlockHash<N>
    ) -> BoxFuture<'b, StorageResult<B>>;

    fn insert_block(
        &self, 
        hash: BlockHash<N>, 
        block: B
    ) -> BoxFuture<'_, StorageResult<()>>;

    fn get_qc<'b, 'a: 'b>(
        &'a self, 
        hash: &'b BlockHash<N>
    ) -> BoxFuture<'b, StorageResult<QuorumCertificate<N>>>;

    fn get_qc_for_view(
        &self, 
        view: u64
    ) -> BoxFuture<'_, StorageResult<QuorumCertificate<N>>>;

    fn insert_qc(
        &self, 
        qc: QuorumCertificate<N>
    ) -> BoxFuture<'_, StorageResult<()>>;

    fn get_leaf<'b, 'a: 'b>(
        &'a self, 
        hash: &'b LeafHash<N>
    ) -> BoxFuture<'b, StorageResult<Leaf<B, N>>>;

    fn get_leaf_by_block<'b, 'a: 'b>(
        &'a self, 
        hash: &'b BlockHash<N>
    ) -> BoxFuture<'b, StorageResult<Leaf<B, N>>>;

    fn insert_leaf(&self, leaf: Leaf<B, N>) -> BoxFuture<'_, StorageResult<()>>;

    fn insert_state(
        &self, 
        state: S, 
        hash: LeafHash<N>
    ) -> BoxFuture<'_, StorageResult<()>>;

    fn get_state<'b, 'a: 'b>(
        &'a self, 
        hash: &'b LeafHash<N>
    ) -> BoxFuture<'_, StorageResult<S>>;
}

Abstraction over on disk persistence of node state;

This should be a cloneable handle to an underlying storage, with each clone pointing to the same underlying storage.

This trait has been constructed for object saftey over convenience.

Required methods

fn get_block<'b, 'a: 'b>(
&'a self,
hash: &'b BlockHash<N>
) -> BoxFuture<'b, StorageResult<B>>

Retrieves a block from storage, returning None if it could not be found in local storage

fn insert_block(
&self,
hash: BlockHash<N>,
block: B
) -> BoxFuture<'_, StorageResult<()>>

Inserts a block into storage

fn get_qc<'b, 'a: 'b>(
&'a self,
hash: &'b BlockHash<N>
) -> BoxFuture<'b, StorageResult<QuorumCertificate<N>>>

Retrieves a Quorum Certificate from storage, by the hash of the block it refers to

fn get_qc_for_view(
&self,
view: u64
) -> BoxFuture<'_, StorageResult<QuorumCertificate<N>>>

Retrieves the Quorum Certificate associated with a particular view number

fn insert_qc(
&self,
qc: QuorumCertificate<N>
) -> BoxFuture<'_, StorageResult<()>>

Inserts a Quorum Certificate into the storage. Should reject the QC if it is malformed or not from a decide stage

fn get_leaf<'b, 'a: 'b>(
&'a self,
hash: &'b LeafHash<N>
) -> BoxFuture<'b, StorageResult<Leaf<B, N>>>

Retrieves a leaf by its hash

fn get_leaf_by_block<'b, 'a: 'b>(
&'a self,
hash: &'b BlockHash<N>
) -> BoxFuture<'b, StorageResult<Leaf<B, N>>>

Retrieves a leaf by the hash of its block

fn insert_leaf(&self, leaf: Leaf<B, N>) -> BoxFuture<'_, StorageResult<()>>

Inserts a leaf

fn insert_state(
&self,
state: S,
hash: LeafHash<N>
) -> BoxFuture<'_, StorageResult<()>>

Inserts a State, indexed by the hash of the Leaf that created it

fn get_state<'b, 'a: 'b>(
&'a self,
hash: &'b LeafHash<N>
) -> BoxFuture<'_, StorageResult<S>>

Retrieves a State, indexed by the hash of the Leaf that created it