Struct phaselock::types::PhaseLockHandle

pub struct PhaseLockHandle<I: NodeImplementation<N> + Send + Sync + 'static, const N: usize> { /* private fields */ }

Event streaming handle for a PhaseLock instance running in the background

This type provides the means to message and interact with a background PhaseLock instance, allowing the ability to receive Events from it, send transactions to it, and interact with the underlying storage.

Implementations

impl<I: NodeImplementation<N> + 'static, const N: usize> PhaseLockHandle<I, N>

pub async fn next_event(
&mut self
) -> Result<Event<I::Block, I::State>, PhaseLockError>

Will return the next event in the queue

Errors

Will return PhaseLockError::NetworkFault if the underlying PhaseLock has been closed.

pub fn next_event_sync(
&mut self
) -> Result<Event<I::Block, I::State>, PhaseLockError>

Syncronous version of next_event

Will internally call block_on on next_event

Errors

See documentation for next_event

pub fn try_next_event(
&mut self
) -> Result<Option<Event<I::Block, I::State>>, PhaseLockError>

Will attempt to immediatly pull an event out of the queue

Errors

Will return PhaseLockError::NetworkFault if the underlying PhaseLock instance has shut down

pub fn availible_events(
&mut self
) -> Result<Vec<Event<I::Block, I::State>>, PhaseLockError>

Will pull all the currently available events out of the event queue.

Errors

Will return PhaseLockError::NetworkFault if the underlying PhaseLock instance has been shut down.

pub async fn get_state(&self) -> Arc<I::State>

Gets the current commited state of the PhaseLock instance

pub fn get_state_sync(&self) -> Arc<I::State>

Gets the current commited state of the PhaseLock instance, blocking on the future

pub async fn submit_transaction(
&self,
tx: <<I as NodeImplementation<N>>::Block as BlockContents<N>>::Transaction
) -> Result<(), PhaseLockError>

Submits a transaction to the backing PhaseLock instance.

The current node broadcasts the transaction to all nodes on the network.

Errors

Will return a PhaseLockError if some error occurs in the underlying PhaseLock instance.

pub fn submit_transaction_sync(
&self,
tx: <<I as NodeImplementation<N>>::Block as BlockContents<N>>::Transaction
) -> Result<(), PhaseLockError>

Sycronously sumbits a transaction to the backing PhaseLock instance.

Errors

See documentation for submit_transaction

pub async fn start(&self)

Signals to the underlying PhaseLock to unpause

This will cause the background task to start running consensus again.

pub fn start_sync(&self)

Synchronously signals the underlying PhaseLock to unpause

pub async fn pause(&self)

Signals the underlying PhaseLock to pause

This will cause the background task to stop driving consensus after the completion of the current view.

pub fn pause_sync(&self)

Synchronously signals the underlying PhaseLock to pause

pub async fn run_one_round(&self)

Signals the underlying PhaseLock to run one round, if paused.

Do not call this function if PhaseLock has been unpaused by PhaseLockHandle::start.

pub fn run_one_round_sync(&self)

Synchronously signals the underlying PhaseLock to run one round, if paused

pub fn storage(&self) -> &I::Storage

Provides a reference to the underlying storage for this PhaseLock, allowing access to historical data

pub async fn shut_down(self)

Shut down the the inner phaselock and wait until all background threads are closed.

pub async fn get_round_runner_state(
&self
) -> Result<RoundRunnerState, Box<dyn Error>>

Get the state of the internal round runner. This is used for testing purposes.

Errors

Will return an error when the background thread is unable to respond within a second.

pub fn get_next_view_timeout(&self) -> u64

return the timeout for a view of the underlying Phaselock