Module phaselock::demos::dentry

Double entry accounting demo

This module provides an implementation of the PhaseLock suite of traits that implements a basic demonstration of double entry accounting.

These implementations are useful in examples and integration testing, but are not suitable for production use.

Structs

Addition

Records an increase in an account balance

DEntryBlock

The block for the dentry demo

DEntryNode

The node implementation for the dentry demo

State

The state for the dentry demo

Subtraction

Records a reduction in an account balance

Transaction

The transaction for the dentry demo

Enums

DEntryError

The error type for the dentry demo

Type Definitions

Account

The account identifier type used by the demo

Balance

The account balance type used by the demo