Enum phaselock::traits::NetworkError

pub enum NetworkError {

    ListenerSend,
    CouldNotDeliver,
    NoSuchNode,
    FailedToSerialize {
        source: Box<ErrorKind, Global>,
    },
    FailedToDeserialize {
        source: Box<ErrorKind, Global>,
    },
    WebSocket {
        source: Error,
    },
    ExecutorError {
        source: Error,
    },
    SocketDecodeError {
        input: String,
        source: Error,
    },
    FailedToBindListener {
        source: Error,
    },
    NoSocketsError {
        input: String,
    },
    Other {
        inner: Box<dyn Error + Send + Sync + 'static, Global>,
    },
    ChannelSend,
    IdentityHandshake,
    ShutDown,

}

Error type for networking

Variants

ListenerSend

A Listener failed to send a message

CouldNotDeliver

Could not deliver a message to a specified recipient

NoSuchNode

Attempted to deliver a message to an unknown node

FailedToSerialize

Fields

source: Box<ErrorKind, Global>

Originating bincode error

Failed to serialize a message

FailedToDeserialize

Fields

source: Box<ErrorKind, Global>

originating bincode error

Failed to deserealize a message

WebSocket

Fields

source: Error

Originating websockets error

WebSockets specific error

ExecutorError

Fields

source: Error

Originating async_std error

Error orginiating from within the executor

SocketDecodeError

Fields

input: String

Input that was given

source: Error

Originating io error

Failed to decode a socket specification

FailedToBindListener

Fields

source: Error

originating io error

Failed to bind a listener socket

NoSocketsError

Fields

input: String

Input that was given

No sockets were open

Other

Fields

inner: Box<dyn Error + Send + Sync + 'static, Global>

Originating error

Generic error type for compatibility if needed

ChannelSend

Channel error

IdentityHandshake

Could not complete handshake

ShutDown

The underlying connection has been shut down