Trait phaselock::traits::NetworkingImplementation

pub trait NetworkingImplementation<M>: Send + Sync where
    M: 'static + Serialize + DeserializeOwned + Send + Clone, {
    fn broadcast_message(
        &self, 
        message: M
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send, Global>>;

    fn message_node(
        &self, 
        message: M, 
        recipient: PubKey
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send, Global>>;

    fn broadcast_queue(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send, Global>>;

    fn next_broadcast(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send, Global>>;

    fn direct_queue(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send, Global>>;

    fn next_direct(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send, Global>>;

    fn known_nodes(
        &self
    ) -> Pin<Box<dyn Future<Output = Vec<PubKey, Global>> + Send, Global>>;

    fn obj_clone(
        &self
    ) -> Box<dyn NetworkingImplementation<M> + 'static, Global>;

    fn network_changes(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<NetworkChange, Global>, NetworkError>> + Send, Global>>;
}

Describes, generically, the behaviors a networking implementation must have

Required methods

fn broadcast_message(
&self,
message: M
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send, Global>>

Broadcasts a message to the network

Should provide that the message eventually reach all non-faulty nodes

fn message_node(
&self,
message: M,
recipient: PubKey
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send, Global>>

Sends a direct message to a specific node

fn broadcast_queue(
&self
) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send, Global>>

Moves out the entire queue of received broadcast messages, should there be any

Provided as a future to allow the backend to do async locking

fn next_broadcast(
&self
) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send, Global>>

Provides a future for the next received broadcast

Will unwrap the underlying NetworkMessage

fn direct_queue(
&self
) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send, Global>>

Moves out the entire queue of received direct messages to this node

fn next_direct(
&self
) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send, Global>>

Provides a future for the next received direct message to this node

Will unwrap the underlying NetworkMessage

fn known_nodes(
&self
) -> Pin<Box<dyn Future<Output = Vec<PubKey, Global>> + Send, Global>>

Node’s currently known to the networking implementation

Kludge function to work around leader election

fn obj_clone(&self) -> Box<dyn NetworkingImplementation<M> + 'static, Global>

Object safe clone

fn network_changes(
&self
) -> Pin<Box<dyn Future<Output = Result<Vec<NetworkChange, Global>, NetworkError>> + Send, Global>>

Returns a list of changes in the network that have been observed. Calling this function will clear the internal list.