Trait phaselock::traits::NodeImplementation

pub trait NodeImplementation<const N: usize>: 'static + Send + Sync + Debug + Clone {
    type Block: 'static + BlockContents<N>;
    type State: State<N>;
    type Storage: Storage<Self::Block, Self::State, N> + Clone;
    type Networking: NetworkingImplementation<Message<Self::Block, <Self::Block as BlockContents<N>>::Transaction, Self::State, N>> + Clone;
    type StatefulHandler: StatefulHandler<N>;
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow PhaseLock to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as PhaseLockdoes not actually store or keep a reference to any value implementing this trait.

Associated Types

type Block: 'static + BlockContents<N>

Block type for this consensus implementation

type State: State<N>

State type for this consensus implementation

type Storage: Storage<Self::Block, Self::State, N> + Clone

Storage type for this consensus implementation

type Networking: NetworkingImplementation<Message<Self::Block, <Self::Block as BlockContents<N>>::Transaction, Self::State, N>> + Clone

Networking type for this consensus implementation

type StatefulHandler: StatefulHandler<N>

Stateful call back handler for this consensus implementation