Trait phaselock::traits::Storage

pub trait Storage<B, S, const N: usize>: Clone + Send + Sync where
    B: 'static + BlockContents<N>,
    S: 'static + State<N, Block = B>, {
    fn get_block<'b, 'a>(
        &'a self, 
        hash: &'b BlockHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<B>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>>
    where
        'a: 'b;

    fn insert_block(
        &self, 
        hash: BlockHash<N>, 
        block: B
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>;

    fn get_qc<'b, 'a>(
        &'a self, 
        hash: &'b BlockHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>>
    where
        'a: 'b;

    fn get_qc_for_view(
        &self, 
        view: u64
    ) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>;

    fn insert_qc(
        &self, 
        qc: QuorumCertificate<N>
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>;

    fn get_leaf<'b, 'a>(
        &'a self, 
        hash: &'b LeafHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>>
    where
        'a: 'b;

    fn get_leaf_by_block<'b, 'a>(
        &'a self, 
        hash: &'b BlockHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>>
    where
        'a: 'b;

    fn insert_leaf(
        &self, 
        leaf: Leaf<B, N>
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>;

    fn insert_state(
        &self, 
        state: S, 
        hash: LeafHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>;

    fn get_state<'b, 'a>(
        &'a self, 
        hash: &'b LeafHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<S>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>>
    where
        'a: 'b;

    fn commit(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>;
}

Abstraction over on disk persistence of node state

This should be a cloneable handle to an underlying storage, with each clone pointing to the same underlying storage.

This trait has been constructed for object saftey over convenience.

Required methods

fn get_block<'b, 'a>(
&'a self,
hash: &'b BlockHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<B>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>> where
'a: 'b,

Retrieves a block from storage, returning None if it could not be found in local storage

fn insert_block(
&self,
hash: BlockHash<N>,
block: B
) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>

Inserts a block into storage. Make sure to call [commit] after all data is inserted.

fn get_qc<'b, 'a>(
&'a self,
hash: &'b BlockHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>> where
'a: 'b,

Retrieves a Quorum Certificate from storage, by the hash of the block it refers to

fn get_qc_for_view(
&self,
view: u64
) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>

Retrieves the Quorum Certificate associated with a particular view number

fn insert_qc(
&self,
qc: QuorumCertificate<N>
) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>

Inserts a Quorum Certificate into the storage. Should reject the QC if it is malformed or not from a decide stage. Make sure to call [commit] after all data is inserted.

fn get_leaf<'b, 'a>(
&'a self,
hash: &'b LeafHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>> where
'a: 'b,

Retrieves a leaf by its hash

fn get_leaf_by_block<'b, 'a>(
&'a self,
hash: &'b BlockHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>> where
'a: 'b,

Retrieves a leaf by the hash of its block

fn insert_leaf(
&self,
leaf: Leaf<B, N>
) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>

Inserts a leaf. Make sure to call [commit] after all data is inserted.

fn insert_state(
&self,
state: S,
hash: LeafHash<N>
) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>

Inserts a State, indexed by the hash of the Leaf that created it. Make sure to call [commit] after all data is inserted.

fn get_state<'b, 'a>(
&'a self,
hash: &'b LeafHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<S>, Box<dyn Error + Send + Sync + 'static, Global>>> + Send + 'b, Global>> where
'a: 'b,

Retrieves a State, indexed by the hash of the Leaf that created it

fn commit(
&self
) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error + Send + Sync + 'static, Global>>> + Send, Global>>

Commit the changes in the Storage. This should be called every time one or multiple of the following functions are called:

insert_block

insert_qc

insert_leaf

insert_state