Trait phaselock::traits::Storage

pub trait Storage<B, S, const N: usize>: Clone + Send + Sync where
    B: 'static + BlockContents<N>,
    S: 'static + State<N, Block = B>, {
    fn get_block<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        hash: &'life1 BlockHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<B>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_qc<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        hash: &'life1 BlockHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_qc_for_view<'life0, 'async_trait>(
        &'life0 self, 
        view: ViewNumber
    ) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_leaf<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        hash: &'life1 LeafHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_leaf_by_block<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        hash: &'life1 BlockHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_state<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        hash: &'life1 LeafHash<N>
    ) -> Pin<Box<dyn Future<Output = Result<Option<S>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn update<'a, 'async_trait, F, FUT>(
        &'a self, 
        update_fn: F
    ) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        F: 'a + FnOnce(Box<dyn StorageUpdater<'a, B, S, N> + 'a, Global>) -> FUT + Send + 'async_trait,
        FUT: 'a + Future<Output = Result<(), StorageError>> + Send + 'async_trait,
        Self: 'async_trait;
    fn get_internal_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = StorageState<B, S, N>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_newest_qc<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Abstraction over on disk persistence of node state

This should be a cloneable handle to an underlying storage, with each clone pointing to the same underlying storage.

This trait has been constructed for object saftey over convenience.

Required methods

fn get_block<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 BlockHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<B>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves a block from storage, returning None if it could not be found in local storage

fn get_qc<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 BlockHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves a Quorum Certificate from storage, by the hash of the block it refers to

fn get_qc_for_view<'life0, 'async_trait>(
&'life0 self,
view: ViewNumber
) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Retrieves the Quorum Certificate associated with a particular view number

fn get_leaf<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 LeafHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves a leaf by its hash

fn get_leaf_by_block<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 BlockHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<Leaf<B, N>>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves a leaf by the hash of its block

fn get_state<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 LeafHash<N>
) -> Pin<Box<dyn Future<Output = Result<Option<S>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves a State, indexed by the hash of the Leaf that created it

fn update<'a, 'async_trait, F, FUT>(
&'a self,
update_fn: F
) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>> where
'a: 'async_trait,
F: 'a + FnOnce(Box<dyn StorageUpdater<'a, B, S, N> + 'a, Global>) -> FUT + Send + 'async_trait,
FUT: 'a + Future<Output = Result<(), StorageError>> + Send + 'async_trait,
Self: 'async_trait,

Calls the given update_fn for a list of modifications, then stores these.

If an error occurs somewhere, the caller can assume that no data is stored at all.

fn get_internal_state<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = StorageState<B, S, N>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the internal state of this storage system.

This function should only be used for testing, never in production code.

fn get_newest_qc<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<QuorumCertificate<N>>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Retrieves the newest Quorum Certificate